+++ packages/@magic-ext/algorand/src/types.ts
@@ -1,0 +1,15 @@
+export interface AlgorandConfig {
+  rpcUrl: string;
+}
+
+export interface ConfigType {
+  rpcUrl: string;
+  chainType: string;
+}
+
+export enum AlgorandPayloadMethod {
+  AlgorandSignTransaction = 'algod_signTransaction',
+  AlgorandSignBid = 'algod_signBid',
+  AlgorandGetWallet = 'algod_getWallet',
+  AlgorandSignGroupTransaction = 'algod_signGroupTransaction',
+}