+++ packages/@magic-ext/flow/src/index.cdn.ts
@@ -1,0 +1,3 @@
+import { FlowExtension } from './index';
+
+export { FlowExtension as default };