+++ packages/@magic-ext/harmony/src/index.cdn.ts
@@ -1,0 +1,3 @@
+import { HarmonyExtension } from './index';
+
+export { HarmonyExtension as default };