+++ packages/@magic-ext/harmony/src/types.ts
@@ -1,0 +1,9 @@
+export enum HarmonyPayloadMethod {
+  HarmonySignTransaction = 'hmy_sendTransaction',
+}
+
+export interface HarmonyConfig {
+  rpcUrl: string;
+  chainId: string;
+  chainType: string;
+}