+++ packages/@magic-ext/icon/src/index.cdn.ts
@@ -1,0 +1,3 @@
+import { IconExtension } from './index';
+
+export { IconExtension as default };