+++ packages/@magic-ext/icon/src/type.ts
@@ -1,0 +1,8 @@
+export interface IconConfig {
+  rpcUrl: string;
+}
+
+export interface ConfigType {
+  rpcUrl: string;
+  chainType: string;
+}