+++ packages/@magic-ext/near/src/index.cdn.ts
@@ -1,0 +1,3 @@
+import { NearExtension } from './index';
+
+export { NearExtension as default };