+++ packages/@magic-ext/polkadot/src/type.ts
@@ -1,0 +1,8 @@
+export interface PolkadotConfig {
+  rpcUrl: string;
+}
+
+export interface ConfigType {
+  rpcUrl: string;
+  chainType: string;
+}