+++ packages/@magic-ext/solana/src/index.cdn.ts
@@ -1,0 +1,3 @@
+import { SolanaExtension } from './index';
+
+export { SolanaExtension as default };