+++ packages/@magic-ext/tezos/src/index.cdn.ts
@@ -1,0 +1,3 @@
+import { TezosExtension } from './index';
+
+export { TezosExtension as default };