+++ packages/@magic-ext/tezos/src/type.ts
@@ -1,0 +1,8 @@
+export interface TezosConfig {
+  rpcUrl: string;
+}
+
+export interface ConfigType {
+  rpcUrl: string;
+  chainType: string;
+}