+++ packages/@magic-ext/zilliqa/src/index.cdn.ts
@@ -1,0 +1,3 @@
+import { ZilliqaExtension } from './index';
+
+export { ZilliqaExtension as default };