+++ packages/@magic-sdk/pnp/src/utils/script-data.ts
@@ -1,15 +1,32 @@
+/* eslint-disable prefer-destructuring */
+
 const allPossiblePNPScripts = document.querySelectorAll('script[data-magic-publishable-api-key]');
 const thisScript = (document.currentScript ??
   allPossiblePNPScripts[allPossiblePNPScripts.length - 1]) as HTMLScriptElement;
 
 export function getScriptData() {
   const src = new URL(thisScript.getAttribute('src')!);
+
+  // Values derived from `data-*` attributes
   const apiKey = thisScript.dataset.magicPublishableApiKey;
-  const debug = !!thisScript.dataset.debug;
+  const locale = thisScript.dataset.locale;
   const redirectURI = getAbsoluteURL(thisScript.dataset.redirectUri);
   const loginURI = getAbsoluteURL(thisScript.dataset.loginUri);
+  const termsOfServiceURI = getAbsoluteURL(thisScript.dataset.termsOfServiceUri);
+  const privacyPolicyURI = getAbsoluteURL(thisScript.dataset.privacyPolicyUri);
+  const debug = !!thisScript.dataset.debug;
 
-  return { script: thisScript, src, apiKey, redirectURI, loginURI, debug };
+  return {
+    script: thisScript,
+    src,
+    apiKey,
+    locale,
+    redirectURI,
+    loginURI,
+    termsOfServiceURI,
+    privacyPolicyURI,
+    debug,
+  };
 }
 
 export type ScriptData = ReturnType<typeof getScriptData>;