+++ packages/@magic-sdk/provider/src/util/index.ts
@@ -2,7 +2,6 @@ export * from './base64-json';
 export * from './events';
 export * from './get-payload-id';
 export * from './promise-tools';
-export * as semver from './semver';
 export * as storage from './storage';
 export * from './type-guards';
 export * from './url';