+++ packages/@magic-sdk/pnp/CHANGELOG.md
@@ -1,3 +1,27 @@
+# v0.6.0 (Mon Nov 08 2021)
+
+#### 🚀 Enhancement
+
+- Add parameters for ToS and privacy links to `@magic-sdk/pnp` [#240](https://github.com/magiclabs/magic-js/pull/240) ([@smithki](https://github.com/smithki))
+
+#### Authors: 1
+
+- Ian K Smith ([@smithki](https://github.com/smithki))
+
+---
+
+# v0.5.0 (Fri Oct 29 2021)
+
+#### 🚀 Enhancement
+
+- Add support for `data-locale` parameter in `@magic-sdk/pnp` [#234](https://github.com/magiclabs/magic-js/pull/234) ([@smithki](https://github.com/smithki))
+
+#### Authors: 1
+
+- Ian K Smith ([@smithki](https://github.com/smithki))
+
+---
+
 # v0.4.0 (Fri Oct 22 2021)
 
 #### 🚀 Enhancement